Refactor: Show Client Charge
