+++ app/src/main/java/org/mifos/selfserviceapp/models/Charge.java
@@ -167,123 +167,4 @@ public class Charge {
     }
 
 
-    public static class Currency {
-        private String code;
-        private String name;
-        private int decimalPlaces;
-        private String displaySymbol;
-        private String nameCode;
-        private String displayLabel;
-
-        public String getCode() {
-            return code;
-        }
-
-        public void setCode(String code) {
-            this.code = code;
-        }
-
-        public String getName() {
-            return name;
-        }
-
-        public void setName(String name) {
-            this.name = name;
-        }
-
-        public int getDecimalPlaces() {
-            return decimalPlaces;
-        }
-
-        public void setDecimalPlaces(int decimalPlaces) {
-            this.decimalPlaces = decimalPlaces;
-        }
-
-        public String getDisplaySymbol() {
-            return displaySymbol;
-        }
-
-        public void setDisplaySymbol(String displaySymbol) {
-            this.displaySymbol = displaySymbol;
-        }
-
-        public String getNameCode() {
-            return nameCode;
-        }
-
-        public void setNameCode(String nameCode) {
-            this.nameCode = nameCode;
-        }
-
-        public String getDisplayLabel() {
-            return displayLabel;
-        }
-
-        public void setDisplayLabel(String displayLabel) {
-            this.displayLabel = displayLabel;
-        }
-
-    }
-
-
-    public static class ChargeTimeType {
-        private int id;
-        private String code;
-        private String value;
-
-        public int getId() {
-            return id;
-        }
-
-        public void setId(int id) {
-            this.id = id;
-        }
-
-        public String getCode() {
-            return code;
-        }
-
-        public void setCode(String code) {
-            this.code = code;
-        }
-
-        public String getValue() {
-            return value;
-        }
-
-        public void setValue(String value) {
-            this.value = value;
-        }
-    }
-
-
-    public static class ChargeCalculationType {
-        private int id;
-        private String code; // example "chargeCalculationType.flat"
-        private String value;
-
-        public int getId() {
-            return id;
-        }
-
-        public void setId(int id) {
-            this.id = id;
-        }
-
-        public String getCode() {
-            return code;
-        }
-
-        public void setCode(String code) {
-            this.code = code;
-        }
-
-        public String getValue() {
-            return value;
-        }
-
-        public void setValue(String value) {
-            this.value = value;
-        }
-    }
 }
