--- app/src/main/res/drawable/circle_on.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="oval">
-        <solid android:color="#008000"/>
-        <stroke android:width="2dp" android:color="#fff" />
-        <size android:width="10dp" android:height="10dp"/>
-</shape>
